+++ packages/garden/src/lib/components/VirtualGarden/VirtualGarden.tsx
@@ -12,12 +12,15 @@ import { GardenItemContainer } from '../GardenItemContainer/GardenItemContainer'
 import { HeaderContainer } from '../HeaderContainer/HeaderContainer';
 import { Layout } from '../Layout/Layout';
 
-type VirtualGardenProps<T> = {
+type VirtualGardenProps<TData> = {
 	width?: number;
-	handleOnItemClick: (item: T) => void;
+	handleOnItemClick: (item: TData) => void;
 };
 
-export const VirtualGarden = <T,>({ width, handleOnItemClick }: VirtualGardenProps<T>): JSX.Element => {
+export const VirtualGarden = <TData, TExtendedFields extends string, TCustomGroupByKeys, TCustomState, TContext>({
+	width,
+	handleOnItemClick,
+}: VirtualGardenProps<TData>): JSX.Element => {
 	const parentRef = useRef<HTMLDivElement | null>(null);
 	const containerRef = useRef<HTMLDivElement | null>(null);
 
@@ -96,15 +99,31 @@ export const VirtualGarden = <T,>({ width, handleOnItemClick }: VirtualGardenPro
 			<HeaderContainer highlightedColumn={highlightedColumn} columnVirtualizer={columnVirtualizer} />
 			{columnVirtualizer.virtualItems.map((virtualColumn) => {
 				const currentColumn = garden[virtualColumn.index];
-				const columnItems = getGardenItems<T>(currentColumn as GardenGroup<T>, true);
+				const columnItems = getGardenItems<TData>(currentColumn as GardenGroup<TData>, true);
 
 				return (
 					<Fragment key={virtualColumn.index}>
 						<GardenItemContainer
 							rowVirtualizer={rowVirtualizer}
 							items={columnItems}
-							packageChild={packageChild as CustomVirtualViews<T>['customItemView']}
-							customSubGroup={customGroupView as CustomVirtualViews<T>['customGroupView']}
+							packageChild={
+								packageChild as CustomVirtualViews<
+									TData,
+									TExtendedFields,
+									TCustomGroupByKeys,
+									TCustomState,
+									TContext
+								>['customItemView']
+							}
+							customSubGroup={
+								customGroupView as CustomVirtualViews<
+									TData,
+									TExtendedFields,
+									TCustomGroupByKeys,
+									TCustomState,
+									TContext
+								>['customGroupView']
+							}
 							handleExpand={handleExpand}
 							itemWidth={width}
 							handleOnClick={handleOnItemClick}