+++ packages/garden/src/lib/utils/index.ts
@@ -3,3 +3,7 @@ export * from './groupBy';
 export * from './defaultItemColor';
 export * from './getDescriptionWidth';
 export * from './isSubGroup';
+export * from './defaultSortFunction';
+export * from './getGardenItems';
+export * from './getRowCount';
+export * from './timeout';