+++ packages/workspace-fusion/src/lib/integrations/status-bar/status-bar.ts
@@ -1,0 +1,10 @@
+/**
+ * Integrations folder for exporting types from the integration
+ * namespace: @equinor/workspace-fusion/status-bar
+ */
+
+import { StatusItem } from '@equinor/workspace-status-bar';
+
+type StatusBarConfig<TData> = (data: TData[]) => [StatusItem, ...StatusItem[]];
+
+export type { StatusItem, StatusBarConfig };