+++ packages/workspace-fusion/src/lib/utils/garden/addGarden.tsx
@@ -1,5 +1,6 @@
 import { Garden, GardenController } from '@equinor/workspace-garden';
 import { WorkspaceViewController } from '@equinor/workspace-react';
+import { GardenConfig } from '../../integrations/garden';
 import { GardenIcon } from '../../icons/GardenIcon';
 import { FusionMediator, WorkspaceTabNames } from '../../types';
 import { GetIdentifier } from '../createFusionWorkspace';
@@ -8,25 +9,34 @@ import { configureClickEvents } from './configureClickEvents';
 import { configureDataChange } from './configureDataChange';
 import { configureGardenHighlightSelection } from './configureHighlight';
 import { GardenWorkspaceHeader } from './gardenWorkspaceHeader';
-import { GardenConfig } from '../../types';
 
 export function addGarden<
 	TData extends Record<PropertyKey, unknown>,
+	TExtendedGardenFields extends string,
 	TCustomGroupByKeys extends Record<PropertyKey, unknown>,
 	TCustomState extends Record<PropertyKey, unknown>,
 	TContext extends Record<PropertyKey, unknown>,
 	TError extends Record<PropertyKey, unknown>
 >(
-	gardenConfig: GardenConfig<TData, TCustomGroupByKeys, TCustomState, TContext>,
+	gardenConfig: GardenConfig<TData, TExtendedGardenFields, TCustomGroupByKeys, TCustomState, TContext>,
 	viewController: WorkspaceViewController<WorkspaceTabNames, TError>,
 	mediator: FusionMediator<TData>,
 	getIdentifier: GetIdentifier<TData>
 ) {
-	const gardenController = new GardenController<TData, TCustomGroupByKeys, TCustomState, TContext>({
-		...gardenConfig,
-		data: [],
-		getIdentifier,
-	});
+	const gardenController = new GardenController<
+		TData,
+		TExtendedGardenFields,
+		TCustomGroupByKeys,
+		TCustomState,
+		TContext
+	>(
+		{
+			...gardenConfig,
+			data: [],
+			getIdentifier,
+		},
+		(destroy) => mediator.onUnMount(destroy)
+	);
 	configureDataChange(gardenController, mediator);
 	configureClickEvents(gardenController, mediator, getIdentifier);
 	configureGardenHighlightSelection(gardenController, mediator);
@@ -38,6 +48,4 @@ export function addGarden<
 		TabIcon: GardenIcon,
 		CustomHeader: () => <GardenWorkspaceHeader controller={gardenController} />,
 	});
-
-	mediator.onUnMount(gardenController.destroy);
 }