+++ packages/sidesheet/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.6](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-sidesheet@0.0.5...@equinor/workspace-sidesheet@0.0.6) (2022-11-10)
+
+**Note:** Version bump only for package @equinor/workspace-sidesheet
+
 ## [0.0.5](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-sidesheet@0.0.3...@equinor/workspace-sidesheet@0.0.5) (2022-11-07)
 
 ### Bug Fixes